Test lookup function and fix display function
Do a quick test on the lookup function that prompts a user to chose a number from the displayed list.

Fixed the display function to receive a pointer to a Cell rather than a pointer to the list. This fixes the issue of displaying the list more than once.
At first I passed a pointer which traversed through the entire list while changing the current pointer to the next element in the list. Now it just uses it per Cell.x